Reorganize folder structure to have a BaseListener
