+++ src/libs/Localize/LocaleListener/index.js
@@ -1,0 +1,11 @@
+import BaseLocaleListener from './BaseLocaleListener';
+
+const listenForLocaleChanges = () => {
+    BaseLocaleListener.connect(() => {});
+};
+
+const LocaleListener = {
+    listenForLocaleChanges,
+};
+
+export default LocaleListener;
