--- src/libs/Localize/index.desktop.js
@@ -1,4 +1,0 @@
-import ELECTRON_EVENTS from '../../../desktop/ELECTRON_EVENTS';
-import BaseLocaleListener from './BaseLocaleListener';
-
-BaseLocaleListener.listenForLocaleChanges(val => window.electron.send(ELECTRON_EVENTS.LOCALE_UPDATED, val));
